Move lock grabbing/releasing logic to PersistentLockManager
